+++ IPX800cs/Parsers/v2/Http/IPX800v2GetAnalogInputHttpResponseParser.cs
@@ -5,11 +5,11 @@ namespace IPX800cs.Parsers.v2.Http
 {
     internal class IPX800v2GetAnalogInputHttpResponseParser : IAnalogInputResponseParser
     {
-        public double ParseResponse(string ipxResponse, int inputNumber)
+        public int ParseResponse(string ipxResponse, int inputNumber)
         {
             XDocument xmlDoc = XDocument.Parse(ipxResponse);
             string value = xmlDoc.Element("response").Elements($"an{inputNumber}").First().Value;
-            return double.Parse(value);
+            return int.Parse(value);
         }
     }
 }
